cloud_storage/requirements.txt
azure-storage-blob==12.10.0
google-cloud-storage==1.38.0
